frontend/pyproject.toml
[tool.poetry]
name = "frontend"
version = "0.1.0"
description = ""
authors = ["Ibrahim Al Ghafri <ibrahim.ghafri@rihal.om>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
streamlit = "^1.41.1"
pillow = "^11.1.0"
matplotlib = "^3.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
